move: determine_zoom to common.py
