+++ ocrd_cis/ocropy/common.py
@@ -10,7 +10,7 @@ from scipy import stats, signal
 from skimage.morphology import medial_axis
 import networkx as nx
 from PIL import Image
-
+from ocrd_models import OcrdExif
 from . import ocrolib
 from .ocrolib import morph, psegutils, sl
 # for decorators (type-checks etc):
@@ -2102,3 +2102,15 @@ def lines2regions(binary, llabels,
     #     rlabels[region_hull] = region
     # DSAVE('rlabels_closed', rlabels)
     return rlabels
+
+def determine_zoom(dpi: float, page_image_info: OcrdExif) -> float:
+    if dpi > 0:
+        zoom = 300.0/dpi
+    elif page_image_info.resolution != 1:
+        dpi = page_image_info.resolution
+        if page_image_info.resolutionUnit == 'cm':
+            dpi *= 2.54
+        zoom = 300.0/dpi
+    else:
+        zoom = 1
+    return zoom
