+++ ocrd_cis/ocropy/resegment.py
@@ -30,6 +30,7 @@ from .common import (
     pil2array,
     odd,
     DSAVE,
+    determine_zoom,
     # binarize,
     check_page,
     check_region,
@@ -129,16 +130,9 @@ class OcropyResegment(Processor):
 
             page_image, page_coords, page_image_info = self.workspace.image_from_page(
                 page, page_id, feature_selector='binarized')
-            if self.parameter['dpi'] > 0:
-                zoom = 300.0/self.parameter['dpi']
-            elif page_image_info.resolution != 1:
-                dpi = page_image_info.resolution
-                if page_image_info.resolutionUnit == 'cm':
-                    dpi *= 2.54
-                self.logger.info('Page "%s" uses %f DPI', page_id, dpi)
-                zoom = 300.0/dpi
-            else:
-                zoom = 1
+
+            zoom = determine_zoom(self.parameter['dpi'], page_image_info)
+            self.logger.info(f"Page '{page_id}' uses {self.parameter['dpi']} DPI")
 
             ignore = (page.get_ImageRegion() +
                       page.get_LineDrawingRegion() +
